+++ src/components/zap-mint/submit-zap.tsx
@@ -35,6 +35,7 @@ import {
 import ZapDustWarningCheckbox from './zap-dust-warning-checkbox'
 import ZapErrorMsg, { ZapTxErrorMsg } from './zap-error-msg'
 import ZapPriceImpactWarningCheckbox from './zap-warning-checkbox'
+import CopyTraceId from '../copy-trace-id'
 
 const LoadingButton = ({
   fetchingZapper,
@@ -285,7 +286,7 @@ const SubmitZapButton = ({
             : !readyToSubmit || loadingTx || validatingTx)
         }
         loading={approving || loadingTx || validatingTx || confirmingApproval}
-        // gas={readyToSubmit ? (gas ? BigInt(gas) : undefined) : approvalGas}
+        gas={readyToSubmit ? (gas ? BigInt(gas) * gasMultiplier : undefined) : undefined}
         onClick={() => {
           setOngoingTx(true)
           if (readyToSubmit) {
@@ -304,6 +305,7 @@ const SubmitZapButton = ({
           ? `${addStepTwoLabel ? 'Step 2. ' : ''}${buttonLabel}`
           : `${addStepOneLabel ? 'Step 1. ' : ''}Approve use of ${inputSymbol}`}
       </TransactionButton>
+      <CopyTraceId enabled={!error} />
       <ZapTxErrorMsg error={error} />
     </div>
   )